requirements.txt
Django==5.0.4
djangorestframework==3.14.0
Pillow==10.2.0
python-decouple==3.8
django-browser-reload==1.12.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0